[update] initialized excel file upload user action
