+++ SPFAdminSystem/SPFAdminSystem/Database/DataContext.cs
@@ -1,4 +1,5 @@
 ﻿using Microsoft.EntityFrameworkCore;
+using SPFAdminSystem.Database.UserFiles;
 using System;
 using System.Collections.Generic;
 
@@ -8,7 +9,7 @@ public class DataContext : DbContext
     public DbSet<Product> Products => Set<Product>();
     public DbSet<UserAction> UserActions => Set<UserAction>();
     public DbSet<Mapping> Mappings => Set<Mapping>();
-
+    public DbSet<UploadMappingFileChangeProduct> UploadMappingFileChangeProducts => Set<UploadMappingFileChangeProduct>();
 
 
     public DataContext(DbContextOptions<DataContext> options) : base(options)
